+++ CHANGES.md
@@ -1,3 +1,4 @@
 ## v0.2
 
-* Multiple low incidence results can be adjusted for prior likelihood using `adjust=True`
+* Multiple low incidence results can be adjusted for prior likelihood using `adjust=True`. This is now the default behavior.
+* Straightlining results can be returned non-normalized using `adjust=False`